+++ src/day15_2.cs
@@ -7,7 +7,7 @@ class Day15_2 {
         int result = 0;
 
         foreach(char c in input.ToCharArray()){
-            result = ((result + (int) c) * 17) % 256;
+            result = (result + c) * 17 % 256;
         }
 
         return result;
